# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexgohost Billing & Client Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hosting/plans:
    get:
      operationId: getHostingPlans
      tags: [hosting]
      summary: List hosting plans
      responses:
        "200":
          description: List of hosting plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostingPlan"
    post:
      operationId: createHostingPlan
      tags: [hosting]
      summary: Create hosting plan (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHostingPlanRequest"
      responses:
        "201":
          description: Created hosting plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostingPlan"
  /hosting/plans/{id}:
    put:
      operationId: updateHostingPlan
      tags: [hosting]
      summary: Update hosting plan (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHostingPlanRequest"
      responses:
        "200":
          description: Updated hosting plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostingPlan"
    delete:
      operationId: deleteHostingPlan
      tags: [hosting]
      summary: Delete hosting plan (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/hosting:
    get:
      operationId: getAllHostingServices
      tags: [hosting]
      summary: List all hosting services (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All hosting services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostingService"
  /admin/hosting/{id}/suspend:
    post:
      operationId: suspendHosting
      tags: [hosting]
      summary: Suspend hosting service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hosting suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostingService"
  /admin/hosting/{id}/terminate:
    post:
      operationId: terminateHosting
      tags: [hosting]
      summary: Terminate hosting service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hosting terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostingService"
  /client/hosting:
    get:
      operationId: getMyHosting
      tags: [hosting]
      summary: Get my hosting services
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My hosting services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostingService"
  /domains:
    get:
      operationId: getMyDomains
      tags: [domains]
      summary: Get my domains
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My domains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"
  /admin/domains:
    get:
      operationId: getAllDomains
      tags: [domains]
      summary: List all domains (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All domains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"
  /admin/domains/pricing:
    get:
      operationId: getDomainPricing
      tags: [domains]
      summary: Get domain pricing (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Domain pricing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DomainPricing"
    post:
      operationId: setDomainPricing
      tags: [domains]
      summary: Set domain pricing (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainPricingRequest"
      responses:
        "200":
          description: Domain pricing updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainPricing"
  /admin/domains/{id}/renew:
    post:
      operationId: renewDomain
      tags: [domains]
      summary: Renew domain (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Domain renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
...
  /client/dashboard:
    get:
      operationId: getClientDashboard
      tags: [admin-dashboard]
      summary: Get client dashboard data
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Client dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDashboardData"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    ClientDashboardData:
      type: object
      properties:
        activeServices:
          type: number
        activeDomains:
          type: number
        unpaidInvoices:
          type: number
        openTickets:
          type: number
        recentInvoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        recentTickets:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
      required:
        - activeServices
        - activeDomains
        - unpaidInvoices
        - openTickets
        - recentInvoices
        - recentTickets
...
    HostingPlan:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        billingCycle:
          type: string
          enum: [monthly, yearly]
        diskSpace:
          type: string
        bandwidth:
          type: string
        emailAccounts:
          type: number
        databases:
          type: number
        subdomains:
          type: number
        ftpAccounts:
          type: number
        isActive:
          type: boolean
        features:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - price
        - billingCycle
        - diskSpace
        - bandwidth
        - isActive
    CreateHostingPlanRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        billingCycle:
          type: string
          enum: [monthly, yearly]
        diskSpace:
          type: string
        bandwidth:
          type: string
        emailAccounts:
          type: number
        databases:
          type: number
        subdomains:
          type: number
        ftpAccounts:
          type: number
        features:
          type: array
          items:
            type: string
      required:
        - name
        - price
        - billingCycle
        - diskSpace
        - bandwidth
    HostingService:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        planId:
          type: string
        planName:
          type: string
        domain:
          type: string
        username:
          type: string
        serverIp:
          type: string
        status:
          type: string
          enum: [active, suspended, terminated, pending]
        startDate:
          type: string
        expiryDate:
          type: string
        diskUsed:
          type: string
        bandwidthUsed:
          type: string
        cpanelUrl:
          type: string
      required:
        - id
        - clientId
        - planId
        - planName
        - status
    Domain:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        name:
          type: string
        tld:
          type: string
        registrationDate:
          type: string
        expiryDate:
          type: string
        status:
          type: string
          enum: [active, expired, pending, transferred]
        autoRenew:
          type: boolean
        nameservers:
          type: array
          items:
            type: string
      required:
        - id
        - clientId
        - name
        - tld
        - status
    DomainPricing:
      type: object
      properties:
        id:
          type: string
        tld:
          type: string
        registrationPrice:
          type: number
        renewalPrice:
          type: number
        transferPrice:
          type: number
      required:
        - id
        - tld
        - registrationPrice
        - renewalPrice
    DomainPricingRequest:
      type: object
      properties:
        tld:
          type: string
        registrationPrice:
          type: number
        renewalPrice:
          type: number
        transferPrice:
          type: number
      required:
        - tld
        - registrationPrice
        - renewalPrice
...
    Invoice:
      type: object
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        amount:
          type: number
        tax:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [unpaid, paid, cancelled, overdue]
        dueDate:
          type: string
        paidDate:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - clientId
        - amount
        - total
        - status
        - dueDate
        - createdAt
    InvoiceItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required:
        - description
        - quantity
        - unitPrice
        - total
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        ticketNumber:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        subject:
          type: string
        status:
          type: string
          enum: [open, closed, pending, answered]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        department:
          type: string
        lastReply:
          type: string
        messagesCount:
          type: number
        createdAt:
          type: string
      required:
        - id
        - ticketNumber
        - clientId
        - subject
        - status
        - priority
        - createdAt
...